+++ arduino/Context.h
@@ -45,12 +45,20 @@ public:
         return _recorder;
     }
 
+    int8_t GetRecordRetainDays() const
+    {
+        return _record_retain_days;
+    }
+
+    void SetRecordRetainDays(uint8_t days);
+
 private:
     Buzzer &_buzzer;
     uint8_t _id = 1;
     uint8_t _key[KEY_SIZE] = { 0xff, 0xff, 0xff, 0xff, 0xff, 0xff };
     uint8_t _timeout_hr = 3;
-    uint8_t _reserved[8];
+    uint8_t _record_retain_days = 1;
+    uint8_t _reserved[7];
     AOP::Recorder _recorder;
 
     friend struct RecorderAccess;