+++ arduino/Shell.cpp
@@ -86,6 +86,8 @@ void Shell::_Process()
         Serial.println(F("rec               List punched cards"));
         Serial.println(F("rec 123           Print punch count for a card"));
         Serial.println(F("recclr 123        Clear card from the record"));
+        Serial.println(F("recdays           How many days to keep the record"));
+        Serial.println(F("recdays 1         Clear record after so many days"));
     } else if (_buffer.startsWith(F("info"))) {
         Serial.println(F("version=1.1.0"));
         Serial.print(F("id="));
@@ -104,6 +106,8 @@ void Shell::_Process()
         Serial.print(F(" x "));
         Serial.print(static_cast<int>(_context.GetRecorder().GetBitsPerRecord()));
         Serial.println(F(" bpr"));
+        Serial.print(F("recdays="));
+        _PrintRecordRetainDays();
     } else if (_buffer.startsWith(F("id "))) {
         _SetId(_buffer.c_str() + 3);
         _PrintId();
@@ -135,6 +139,10 @@ void Shell::_Process()
         _RecorderFormat(_buffer.c_str() + 7);
     } else if (_buffer.startsWith(F("recclr "))) {
         _RecorderClear(_buffer.c_str() + 7);
+    } else if (_buffer.startsWith(F("recdays "))) {
+        _SetRecordRetainDays(_buffer.c_str() + 9);
+    } else if (_buffer.startsWith(F("recdays"))) {
+        _PrintRecordRetainDays();
     } else if (_buffer.startsWith(F("rec "))) {
         _RecorderCheck(_buffer.c_str() + 4);
     } else if (_buffer.startsWith(F("rec"))) {
@@ -273,7 +281,7 @@ void Shell::_RecorderFormat(const char *str)
     uint16_t count = ParseNum<uint16_t>(str);
     uint8_t bits_per_record = ParseNum<uint8_t>(str);
 
-    auto res = _context.GetRecorder().Format(count, bits_per_record);
+    auto res = _context.GetRecorder().Format(count, bits_per_record, _context.GetDateTime().unixtime());
     if (res < 0) {
         Serial.print(F("Error "));
         Serial.println(-res);
@@ -316,4 +324,14 @@ void Shell::_RecorderList()
     } printer;
     _context.GetRecorder().List(printer, &printer);
     Serial.println();
+}
+
+void Shell::_PrintRecordRetainDays()
+{
+    Serial.println(_context.GetRecordRetainDays());
+}
+
+void Shell::_SetRecordRetainDays(const char *str)
+{
+    _context.SetRecordRetainDays(ParseNum<uint8_t>(str));
 }